+++ SistemaGestionData/SoldProductData.cs
@@ -54,6 +54,36 @@ namespace SistemaGestionData
                 throw;
             }
         }
+        public static void CreateSoldProducts(Venta sale)
+        {
+            try
+            {
+                using (var context = new ApplicationDbContext())
+                {
+                    if (sale.Productos != null && sale.Productos.Any())
+                    {
+                        foreach (var producto in sale.Productos)
+                        {
+                            var existingProduct = context.Productos.FirstOrDefault(p => p.Descripciones == producto.Descripciones);
+                            var productoVendido = new ProductoVendido
+                            {
+                                IdProducto = existingProduct.Id,
+                                Stock = existingProduct.Stock,
+                                IdVenta = sale.Id
+                            };
+
+                            context.ProductosVendidos.Add(productoVendido);
+                        }
+
+                        context.SaveChanges();
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                throw;
+            }
+        }
         public static void UpdateSoldProduct(ProductoVendido soldProduct)
         {
             try
@@ -70,15 +100,41 @@ namespace SistemaGestionData
                 throw;
             }
         }
-        public static void DeleteSoldProduct(ProductoVendido soldProduct)
+        public static void DeleteSoldProduct(int soldProductId)
         {
             try
             {
                 using (var context = new ApplicationDbContext())
                 {
-                    context.ProductosVendidos.Remove(soldProduct);
-                    context.SaveChanges();
+                    var soldProduct = GetSoldProductById(soldProductId, context);
+
+                    if (soldProduct != null)
+                    {
+                        context.ProductosVendidos.Remove(soldProduct);
+                        context.SaveChanges();
+                    }
                 }
+            }
+            catch (Exception ex)
+            {
+                throw;
+            }
+        }
+        public static void DeleteSoldProducts(int productId)
+        {
+            try
+            {
+                using (var context = new ApplicationDbContext())
+                {
+                    var soldProducts = context.ProductosVendidos.Where(x => x.IdProducto == productId).ToList();
+
+                    if (soldProducts != null && soldProducts.Any())
+                    {
+                        context.ProductosVendidos.RemoveRange(soldProducts);
+                        context.SaveChanges();
+                    }
+                }
+
             }
             catch (Exception ex)
             {
